+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - [Multiple] Better version specific relationships at install and upgrade (#4023 by: HebaruSan)
 - [GUI] Proportional, granular progress updates for installing (#4055 by: HebaruSan)
 - [GUI] Modpack compatibility prompt, GameComparator clean-up (#4056 by: HebaruSan)
+- [ConsoleUI] Add downloads column for ConsoleUI (#4063 by: HebaruSan)
 
 ### Bugfixes